requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
apache-airflow==2.9.1
pandas==2.2.2
faker==25.2.0
pytest==8.2.0
python-dateutil==2.9.0
